+++ src/network/protocol/packet/ContainerOpenPacket.php
@@ -20,7 +20,7 @@ class ContainerOpenPacket extends RakNetDataPacket{
 		$this->reset();
 		$this->putByte($this->windowid);
 		$this->putByte($this->type);
-		$this->putShort($this->slots);
+		$this->putByte($this->slots);
 		$this->putInt($this->x);
 		$this->putInt($this->y);
 		$this->putInt($this->z);
